Changed timer_start to timer_begin, added more performance tests, added functions to estimate number of SPH nodes for cylindrical distributions, added git hash, spheral version, and git branch to adiak data

+++ tests/functional/Strength/TaylorImpact/TaylorImpact.py
@@ -159,21 +159,30 @@ if asph:
     hydroname = "A" + hydroname
 
 # Restart and output files.
-dataDir = os.path.join(baseDir,
-                       geometry,
-                       hydroname,
-                       "XSPH=%s" % XSPH,
-                       "reflect=%s" % reflect,
-                       "%ix%i" % (nr, nz),
-                       "procs=%i" % mpi.procs)
-restartDir = os.path.join(dataDir, "restarts", "proc-%04i" % mpi.rank)
-vizDir = os.path.join(dataDir, "viz")
-restartBaseName = os.path.join(restartDir, "TaylorImpact-%i-%i" % (nr, nz))
+if baseDir:
+    dataDir = os.path.join(baseDir,
+                           geometry,
+                           hydroname,
+                           "XSPH=%s" % XSPH,
+                           "reflect=%s" % reflect,
+                           "%ix%i" % (nr, nz),
+                           "procs=%i" % mpi.procs)
+    restartDir = os.path.join(dataDir, "restarts", "proc-%04i" % mpi.rank)
+    vizDir = os.path.join(dataDir, "viz")
+    restartBaseName = os.path.join(restartDir, "TaylorImpact-%i-%i" % (nr, nz))
+else:
+    dataDir = None
+    restartBaseName = None
+    vizDir = None
+if vizTime is None and vizCycle is None:
+    vizBaseName = None
+else:
+    vizBaseName = "TaylorImpact"
 
 #-------------------------------------------------------------------------------
 # Check if the necessary output directories exist.  If not, create them.
 #-------------------------------------------------------------------------------
-if mpi.rank == 0:
+if mpi.rank == 0 and dataDir:
     if clearDirectories and os.path.exists(dataDir):
         shutil.rmtree(dataDir)
     if not os.path.exists(dataDir):
@@ -182,9 +191,6 @@ if mpi.rank == 0:
         os.makedirs(vizDir)
     if not os.path.exists(restartDir):
         os.makedirs(restartDir)
-mpi.barrier()
-if not os.path.exists(restartDir):
-    os.makedirs(restartDir)
 mpi.barrier()
 
 #-------------------------------------------------------------------------------
@@ -535,7 +541,7 @@ control = SpheralController(integrator, WT,
                             redistributeStep = redistributeStep,
                             restartBaseName = restartBaseName,
                             restoreCycle = restoreCycle,
-                            vizBaseName = "TaylorImpact",
+                            vizBaseName = vizBaseName,
                             vizDir = vizDir,
                             vizStep = vizCycle,
                             vizTime = vizTime)
